requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
Django==3.0
django-jsonfield==1.4.1
django-stripe-payments==2.0.0
djangorestframework==3.14.0
idna==3.4
Pillow==8.4.0
psycopg2-binary==2.8.6
python-decouple==3.7
pytz==2022.7.1
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
stripe==5.1.1
urllib3==1.26.14
